requirements.txt
fastapi
uvicorn
pandas
numpy
requests
openai
python-dotenv
python-multipart
